requirements.txt
fastapi
uvicorn
motor
pymongo
yfinance
openai
requests
python-dotenv
apscheduler
transformers
torch
newspaper3k
beautifulsoup4
pandas
numpy
feedparser
asyncpraw
nest-asyncio
apscheduler